Add 'mergeChanges' parameter to BaseBasicScheduler
This parameter takes a callable, used to tell how incoming changes
should be merged to form a new build request.

This callback takes a list of Change objects. It returns a list of
change IDs: each ID defines a build request which will include this ID
and previous ones. Remaining IDs (after the last mentionned ID, or all
if the returned list is empty) will form an additional build request.